backend/requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
yt-dlp
assemblyai
google-generativeai
python-dotenv
